fix: custom svg for rects (#196)
SDWebImage does not handle SVG very well. A custom parser is needed as all the ones out there are now deprecated for Swift. This adds a simple parser for Svg Rects that allow a user to create custom mini bar charts. It uses the built in XML parser from Apple for iOS

Android Glide does not handle SVG properly either. So AndroidSVG is used to render.

Co-authored-by: Vittorio Cellucci <vel@qlik.com>